requirements-local.txt
autoflake==1.7.7
black==22.10.0
flake8==5.0.4
isort==5.12.0
mypy==1.6.1
pre-commit==2.20.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pyyaml==5.4.1
setuptools==65.4.1
wheel==0.37.1
